Add metrics param for cpu userbenchmark
