+++ userbenchmark/cpu/README.md
@@ -26,9 +26,12 @@ All parameters of `cpu` userbenchmark as below,
   means run eith mdoel predifined default batch size.
 - `--jit` whether to convert and run the model with `jit` mode.
 - `--config, -c` YAML config to specify tests to run.
+- `--metrics` benchmark metrics, split by comma. Current support metrics
+  including `latencies` and `cpu_peak_mem`, default value is `latencies`.
 - `--output, -o` output dir. By default will create folder under
   `.userbenchmark/cpu`.
-- `--timeout` limit single model test run time. Default `None` means no limitation.
+- `--timeout` limit single model test run time. Default `None` means no
+  limitation.
 - `--launcher` whether to use `torch.backends.xeon.run_cpu` to get the peak
   performance on Intel(R) Xeon(R) Scalable Processors.
 - `--launcher-args` work with `--launcher` enabled, to provide the args of
